[JENKINS-23926] - Gracefully integrate changes with #29
