+++ src/main/java/org/jenkinsci/plugins/ownership/model/runs/RunOwnershipAction.java
@@ -26,10 +26,14 @@ package org.jenkinsci.plugins.ownership.model.runs;
 
 import com.synopsys.arc.jenkins.plugins.ownership.IOwnershipHelper;
 import com.synopsys.arc.jenkins.plugins.ownership.ItemOwnershipAction;
-import com.synopsys.arc.jenkins.plugins.ownership.OwnershipDescription;
 import com.synopsys.arc.jenkins.plugins.ownership.OwnershipPlugin;
 import com.synopsys.arc.jenkins.plugins.ownership.util.ui.OwnershipLayoutFormatter;
+import com.synopsys.arc.jenkins.plugins.ownership.wrappers.OwnershipBuildWrapper;;
+import hudson.EnvVars;
+import hudson.model.AbstractBuild;
+import hudson.model.EnvironmentContributingAction;
 import hudson.model.Job;
+import hudson.model.Project;
 import hudson.model.Run;
 import hudson.security.Permission;
 import javax.annotation.Nonnull;
@@ -41,7 +45,8 @@ import javax.annotation.Nonnull;
  * @author Oleg Nenashev <o.v.nenashev@gmail.com>
  * @since 0.6
  */
-public class RunOwnershipAction extends ItemOwnershipAction<Run>  {
+public class RunOwnershipAction extends ItemOwnershipAction<Run> 
+         implements EnvironmentContributingAction {
 
     public RunOwnershipAction(@Nonnull Run describedItem) {
         super(describedItem);
@@ -70,4 +75,18 @@ public class RunOwnershipAction extends ItemOwnershipAction<Run>  {
         return OwnershipPlugin.getInstance().getOwnershipLayoutFormatterProvider().getLayoutFormatter(getDescribedItem());
     }   
     
+    public void buildEnvVars(AbstractBuild<?, ?> build, EnvVars env) {
+        final Job parent = build.getParent();
+        if (!(parent instanceof Project)) {
+            return; // TODO: do something?
+        }
+
+        final Project prj = (Project) parent;
+        final OwnershipBuildWrapper wrapper = (OwnershipBuildWrapper) 
+                prj.getBuildWrappersList().get(OwnershipBuildWrapper.class);
+        if (wrapper == null) {
+            return; // disabled
+        }
+        wrapper.setUp(build, env, null);
+    }
 }
